Place anchor elements in paragraphs
Previously, the anchor elements in index.html were not
contained within paragraph elements. This meant that
the links appeared on the same line on the webpage.

Now, the anchor elements are located within paragraph
elements. This means that the links to the various recipes
now appear on different lines rather than the same line.